+++ crates/miden-objects/src/transaction/proven_tx.rs
@@ -256,6 +256,9 @@ pub struct ProvenTransactionBuilder {
     /// The commitment of the account after the transaction was executed.
     final_account_commitment: Digest,
 
+    /// The commitment of the account delta produced by the transaction.
+    account_delta_commitment: Digest,
+
     /// State changes to the account due to the transaction.
     account_update_details: AccountUpdateDetails,
 
@@ -287,6 +290,7 @@ impl ProvenTransactionBuilder {
         account_id: AccountId,
         initial_account_commitment: Digest,
         final_account_commitment: Digest,
+        account_delta_commitment: Digest,
         ref_block_num: BlockNumber,
         ref_block_commitment: Digest,
         expiration_block_num: BlockNumber,
@@ -296,6 +300,7 @@ impl ProvenTransactionBuilder {
             account_id,
             initial_account_commitment,
             final_account_commitment,
+            account_delta_commitment,
             account_update_details: AccountUpdateDetails::Private,
             input_notes: Vec::new(),
             output_notes: Vec::new(),
@@ -375,6 +380,7 @@ impl ProvenTransactionBuilder {
             self.account_id,
             self.initial_account_commitment,
             self.final_account_commitment,
+            self.account_delta_commitment,
             self.account_update_details,
         );
 
@@ -402,13 +408,16 @@ pub struct TxAccountUpdate {
     /// ID of the account updated by a transaction.
     account_id: AccountId,
 
-    /// The commitment of the account before a transaction was executed.
+    /// The commitment of the account before the transaction was executed.
     ///
     /// Set to `Digest::default()` for new accounts.
     init_state_commitment: Digest,
 
-    /// The commitment of the account state after a transaction was executed.
+    /// The commitment of the account state after the transaction was executed.
     final_state_commitment: Digest,
+
+    /// The commitment to the account delta resulting from the execution of the transaction.
+    account_delta_commitment: Digest,
 
     /// A set of changes which can be applied the account's state prior to the transaction to
     /// get the account state after the transaction. For private accounts this is set to
@@ -422,12 +431,14 @@ impl TxAccountUpdate {
         account_id: AccountId,
         init_state_commitment: Digest,
         final_state_commitment: Digest,
+        account_delta_commitment: Digest,
         details: AccountUpdateDetails,
     ) -> Self {
         Self {
             account_id,
             init_state_commitment,
             final_state_commitment,
+            account_delta_commitment,
             details,
         }
     }
@@ -437,14 +448,19 @@ impl TxAccountUpdate {
         self.account_id
     }
 
-    /// Returns the commitment of the account's initial state.
+    /// Returns the commitment of the account before the transaction was executed.
     pub fn initial_state_commitment(&self) -> Digest {
         self.init_state_commitment
     }
 
-    /// Returns the commitment of the account's after a transaction was executed.
+    /// Returns the commitment of the account after the transaction was executed.
     pub fn final_state_commitment(&self) -> Digest {
         self.final_state_commitment
+    }
+
+    /// Returns the commitment to the account delta resulting from the execution of the transaction.
+    pub fn account_delta_commitment(&self) -> Digest {
+        self.account_delta_commitment
     }
 
     /// Returns the description of the updates for public accounts.
@@ -481,6 +497,7 @@ impl Serializable for TxAccountUpdate {
         self.account_id.write_into(target);
         self.init_state_commitment.write_into(target);
         self.final_state_commitment.write_into(target);
+        self.account_delta_commitment.write_into(target);
         self.details.write_into(target);
     }
 }
@@ -491,6 +508,7 @@ impl Deserializable for TxAccountUpdate {
             account_id: AccountId::read_from(source)?,
             init_state_commitment: Digest::read_from(source)?,
             final_state_commitment: Digest::read_from(source)?,
+            account_delta_commitment: Digest::read_from(source)?,
             details: AccountUpdateDetails::read_from(source)?,
         })
     }
@@ -652,6 +670,7 @@ mod tests {
             AccountId::try_from(ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE).unwrap(),
             Digest::new(EMPTY_WORD),
             Digest::new(EMPTY_WORD),
+            Digest::new(EMPTY_WORD),
             details,
         )
         .validate()
@@ -682,6 +701,7 @@ mod tests {
             AccountId::try_from(ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE).unwrap(),
             Digest::new(EMPTY_WORD),
             Digest::new(EMPTY_WORD),
+            Digest::new(EMPTY_WORD),
             details,
         )
         .validate()
@@ -704,6 +724,8 @@ mod tests {
             [2; 32].try_into().expect("failed to create initial account commitment");
         let final_account_commitment =
             [3; 32].try_into().expect("failed to create final account commitment");
+        let account_delta_commitment =
+            [4; 32].try_into().expect("failed to create account delta commitment");
         let ref_block_num = BlockNumber::from(1);
         let ref_block_commitment = Digest::default();
         let expiration_block_num = BlockNumber::from(2);
@@ -713,6 +735,7 @@ mod tests {
             account_id,
             initial_account_commitment,
             final_account_commitment,
+            account_delta_commitment,
             ref_block_num,
             ref_block_commitment,
             expiration_block_num,