+++ crates/miden-testing/src/kernel_tests/tx/test_epilogue.rs
@@ -12,7 +12,7 @@ use miden_lib::{
     },
 };
 use miden_objects::{
-    account::Account,
+    account::{Account, AccountDelta, AccountStorageDelta, AccountVaultDelta},
     note::{NoteTag, NoteType},
     testing::account_id::ACCOUNT_ID_REGULAR_PRIVATE_ACCOUNT_UPDATABLE_CODE,
     transaction::{OutputNote, OutputNotes},
@@ -81,9 +81,21 @@ fn test_epilogue() {
     )
     .unwrap();
 
-    let mut expected_stack = Vec::with_capacity(16);
+    let account_delta_commitment = AccountDelta::new(
+        tx_context.account().id(),
+        AccountStorageDelta::default(),
+        AccountVaultDelta::default(),
+        ONE,
+    )
+    .unwrap()
+    .commitment();
+
+    let account_update_commitment =
+        miden_objects::Hasher::merge(&[final_account.commitment(), account_delta_commitment]);
+
+    let mut expected_stack = Vec::with_capacity(17);
     expected_stack.extend(output_notes.commitment().as_elements().iter().rev());
-    expected_stack.extend(final_account.commitment().as_elements().iter().rev());
+    expected_stack.extend(account_update_commitment.as_elements().iter().rev());
     expected_stack.push(Felt::from(u32::MAX)); // Value for tx expiration block number
     expected_stack.extend((9..16).map(|_| ZERO));