+++ crates/miden-testing/src/mock_chain/proven_tx_ext.rs
@@ -35,6 +35,7 @@ impl ProvenTransactionExt for ProvenTransaction {
             executed_tx.account_id(),
             executed_tx.initial_account().init_commitment(),
             executed_tx.final_account().commitment(),
+            executed_tx.account_delta().commitment(),
             block_reference.block_num(),
             block_reference.commitment(),
             executed_tx.expiration_block_num(),